backend/requirements.txt
# Backend
fastapi>=0.95.0
uvicorn>=0.21.1
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0
pymongo>=4.5.0,<5.0.0
motor>=3.3.1,<4.0.0
pydantic>=1.10.7,<2.0.0
python-dateutil>=2.8.2,<3.0.0
email-validator>=1.3.1,<2.0.0